+++ aider/coders/base_prompts.py
@@ -14,6 +14,9 @@ You NEVER leave comments describing code without implementing it!
 You always COMPLETELY IMPLEMENT the needed code!
 """
 
+    overeager_prompt = """Pay careful attention to the scope of the user's request.
+Do what they ask, but no more."""
+
     example_messages = []
 
     files_content_prefix = """I have *added these files to the chat* so you can go ahead and edit them.